Return ok body in indexd healthz route
